pyproject.toml
[tool.poetry]
name = "pyserde-timedelta"
version = "0.1.0"
description = "pyserde extension for (de)serializing timedelta"
authors = ["yukinarit <yukinarit84@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yukinarit/pyserde-timedelta"
homepage = "https://github.com/yukinarit/pyserde-timedelta"
packages = [{include = "serde_timedelta"}]
classifiers=[
        "Development Status :: 5 - Production/Stable",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: MIT License",
        "Programming Language :: Python :: 3.8",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
        "Programming Language :: Python :: Implementation :: CPython",
		"Programming Language :: Python :: Implementation :: PyPy",
    ]

[tool.poetry.dependencies]
python = "^3.8"
isodate = "^0.6"
plum-dispatch = "^2.2"
pyserde = ">=0.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
